# openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "BillingInterval": {
        "enum": [
          "monthly",
          "yearly"
        ],
        "title": "BillingInterval",
        "type": "string"
      },
...
      "CheckoutSessionResponse": {
        "properties": {
          "billing_interval": {
            "title": "Billing Interval",
            "type": "string"
          },
          "checkout_url": {
            "title": "Checkout Url",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "checkout_url",
          "tier",
          "billing_interval"
        ],
        "title": "CheckoutSessionResponse",
        "type": "object"
      },
...
      "CreateCheckoutSessionRequest": {
        "properties": {
          "billing_interval": {
            "$ref": "#/components/schemas/BillingInterval",
            "default": "monthly"
          },
          "tier": {
            "enum": [
              "pro",
              "premium"
            ],
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "tier"
        ],
        "title": "CreateCheckoutSessionRequest",
        "type": "object"
      },
...
      "CreatePortalSessionRequest": {
        "properties": {
          "return_path": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "/app/settings/billing",
            "title": "Return Path"
          }
        },
        "title": "CreatePortalSessionRequest",
        "type": "object"
      },
...
      "DailyPickForecast": {
        "additionalProperties": true,
        "description": "Known keys produced by the forecaster for daily pick forecasts.",
        "properties": {
          "analog_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analog Count"
          },
          "expected_return_median_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Return Median Pct"
          },
          "positive_outcome_rate_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Positive Outcome Rate Pct"
          }
        },
        "title": "DailyPickForecast",
        "type": "object"
      },
      "DailyPickItemResponse": {
        "properties": {
          "close_price": {
            "title": "Close Price",
            "type": "number"
          },
          "config_snapshot": {
            "additionalProperties": true,
            "title": "Config Snapshot",
            "type": "object"
          },
          "forecast": {
            "$ref": "#/components/schemas/DailyPickForecast"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "regime_labels": {
            "items": {
              "type": "string"
            },
            "title": "Regime Labels",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "strategy_type": {
            "title": "Strategy Type",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/DailyPickSummary"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "target_dte": {
            "title": "Target Dte",
            "type": "integer"
          }
        },
        "required": [
          "rank",
          "score",
          "symbol",
          "strategy_type",
          "close_price",
          "target_dte"
        ],
        "title": "DailyPickItemResponse",
        "type": "object"
      },
      "DailyPickSummary": {
        "additionalProperties": true,
        "description": "Known keys produced by the backtest engine for daily pick summaries.",
        "properties": {
          "expectancy": {
            "default": 0.0,
            "title": "Expectancy",
            "type": "number"
          },
          "max_drawdown_pct": {
            "default": 0.0,
            "title": "Max Drawdown Pct",
            "type": "number"
          },
          "profit_factor": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profit Factor"
          },
          "sharpe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe Ratio"
          },
          "total_net_pnl": {
            "default": 0.0,
            "title": "Total Net Pnl",
            "type": "number"
          },
          "total_roi_pct": {
            "default": 0.0,
            "title": "Total Roi Pct",
            "type": "number"
          },
          "trade_count": {
            "default": 0,
            "title": "Trade Count",
            "type": "integer"
          },
          "win_rate": {
            "default": 0.0,
            "title": "Win Rate",
            "type": "number"
          }
        },
        "title": "DailyPickSummary",
        "type": "object"
      },
      "DailyPicksResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DailyPickItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "pipeline_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Run Id"
          },
          "pipeline_stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineStatsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trade_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Date"
          }
        },
        "required": [
          "status"
        ],
        "title": "DailyPicksResponse",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "PipelineHistoryItemResponse": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "recommendations_produced": {
            "title": "Recommendations Produced",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "symbols_screened": {
            "title": "Symbols Screened",
            "type": "integer"
          },
          "trade_date": {
            "title": "Trade Date",
            "type": "string"
          }
        },
        "required": [
          "id",
          "trade_date",
          "status",
          "symbols_screened",
          "recommendations_produced"
        ],
        "title": "PipelineHistoryItemResponse",
        "type": "object"
      },
      "PipelineHistoryResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PipelineHistoryItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "PipelineHistoryResponse",
        "type": "object"
      },
      "PipelineStatsResponse": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "full_backtests_run": {
            "title": "Full Backtests Run",
            "type": "integer"
          },
          "pairs_generated": {
            "title": "Pairs Generated",
            "type": "integer"
          },
          "quick_backtests_run": {
            "title": "Quick Backtests Run",
            "type": "integer"
          },
          "recommendations_produced": {
            "title": "Recommendations Produced",
            "type": "integer"
          },
          "symbols_after_screen": {
            "title": "Symbols After Screen",
            "type": "integer"
          },
          "symbols_screened": {
            "title": "Symbols Screened",
            "type": "integer"
          }
        },
        "required": [
          "symbols_screened",
          "symbols_after_screen",
          "pairs_generated",
          "quick_backtests_run",
          "full_backtests_run",
          "recommendations_produced"
        ],
        "title": "PipelineStatsResponse",
        "type": "object"
      },
...
      "PortalSessionResponse": {
        "properties": {
          "portal_url": {
            "title": "Portal Url",
            "type": "string"
          }
        },
        "required": [
          "portal_url"
        ],
        "title": "PortalSessionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "Clerk-issued JWT via Authorization header.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "BacktestForecast API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/billing/checkout-session": {
      "post": {
        "operationId": "create_checkout_session_v1_billing_checkout_session_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Checkout Session",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/portal-session": {
      "post": {
        "operationId": "create_portal_session_v1_billing_portal_session_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePortalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Portal Session",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/webhook": {
      "post": {
        "operationId": "stripe_webhook_v1_billing_webhook_post",
        "parameters": [
          {
            "in": "header",
            "name": "Stripe-Signature",
            "required": true,
            "schema": {
              "title": "Stripe-Signature",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "contentMediaType": "application/octet-stream",
                "maxLength": 65536,
                "title": "Payload",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Stripe Webhook V1 Billing Webhook Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stripe Webhook",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/daily-picks": {
      "get": {
        "description": "Return the latest daily recommendations.\n\nPro+ feature gated via ensure_forecasting_access.",
        "operationId": "get_latest_daily_picks_v1_daily_picks_get",
        "parameters": [
          {
            "in": "query",
            "name": "trade_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trade Date"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyPicksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Latest Daily Picks",
        "tags": [
          "daily-picks"
        ]
      }
    },
    "/v1/daily-picks/history": {
      "get": {
        "description": "Return recent pipeline run history (Pro+ gated).\n\nPipeline runs are system-wide (not user-scoped) since they represent\nshared nightly scanning results.  Access is gated by forecasting\nentitlement.\n\nSupports optional cursor-based pagination via the ``cursor`` parameter\nwhich should be the ``created_at`` ISO timestamp of the last item from\nthe previous page.",
        "operationId": "get_pipeline_history_v1_daily_picks_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 30,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "created_at ISO cursor from previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "created_at ISO cursor from previous page",
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pipeline History",
        "tags": [
          "daily-picks"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}